requirements.txt
Click==7.0
dominate==2.3.5
Flask==1.0.2
Flask-Bootstrap==3.3.7.1
Flask-Moment==0.6.0
flask-paginate==0.5.2
influxdb==5.2.2
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
psutil==5.6.1
python-dotenv==0.10.1
PyYAML==3.12
urllib3==1.24.1
visitor==0.1.3
Werkzeug==0.14.1
